backend/requirements.txt
boto3
passlib
python-jose
pydantic[email]